Add code to save and delete images in Documents
Restructured the Photo entity to store a string identifier and then source image data from the Documents directory. Image data is no longer stored within the persistent store. Added a new class called "ImageHandler" which will feature the code to obtain the URL to the Documents directory and store image data to said directory. Removed the ImageCache class as image data is now stored in Documents directory and is automatically deleted when associated Photo entity is deleted from the persistent store.